Add API endpoints for all user's posts
